requirements.txt
Flask==2.2.3
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pandas==2.0.0
scikit-learn==1.2.2
spacy==3.5.3
nltk==3.8.1
reportlab==3.6.12
fpdf==1.7.2
flask-cors==3.0.10
matplotlib==3.7.1
numpy==1.24.3
Werkzeug==2.2.3
Jinja2==3.1.2
alembic==1.10.4
SQLAlchemy==2.0.9